requirements.txt
fastapi==0.112.0
uvicorn==0.30.3
duckdb==1.0.0
pandas==2.2.2
openpyxl==3.1.5
python-multipart==0.0.9
python-pptx==0.6.23